RES: Optimize memory consumption of `Map<K, SmartList<V>>`
