+++ src/main/kotlin/org/rust/lang/core/resolve2/util/PerNsHashMap.kt
@@ -7,15 +7,12 @@ package org.rust.lang.core.resolve2.util
 
 import com.intellij.util.ArrayUtil.EMPTY_BYTE_ARRAY
 import gnu.trove.THash
-import gnu.trove.THashMap
-import gnu.trove.TObjectHash
 import org.rust.lang.core.resolve.Namespace
 import org.rust.lang.core.resolve2.*
-import java.util.*
 
 /**
  * This is memory-optimized implementation of `HashMap<String, PerNs>`.
- * On real projects only few percents of [PerNs] has more then one [VisItem].
+ * On real projects only few percents of [PerNs] has more than one [VisItem].
  * So almost all [PerNs] has exactly one [VisItem],
  * and we can try to use `HashMap<String, Any>`, where value is either [PerNs] or [VisItem].
  * Here we go even further, and inline [VisItem] right into [items] array.
@@ -25,10 +22,10 @@ import java.util.*
 class PerNsHashMap<K : Any>(
     private val containingMod: ModData,
     private val rootMod: ModData,
-) : PerNsHashMapBase<K, PerNs>() {
+) : THashMapBase<K, PerNs>() {
 
     /**
-     * Each value is either [PerNs] (when it has more then one [VisItem]),
+     * Each value is either [PerNs] (when it has more than one [VisItem]),
      * or [ModPath] corresponding to [VisItem.path] (in this case other [VisItem] fields are stored in [masks])
      */
     private var items: Array<Any? /* ModPath or PerNs */> = THash.EMPTY_OBJECT_ARRAY
@@ -42,15 +39,6 @@ class PerNsHashMap<K : Any>(
      * See [encodeMask] and `decode*`
      */
     private var masks: ByteArray = EMPTY_BYTE_ARRAY
-
-    override fun setUp(initialCapacity: Int): Int {
-        val capacity = super.setUp(initialCapacity)
-        if (initialCapacity != THash.JUST_CREATED_CAPACITY) {
-            items = arrayOfNulls(capacity)
-            masks = ByteArray(capacity)
-        }
-        return capacity
-    }
 
     override fun getValueAtIndex(index: Int): PerNs? {
         val path = items[index] ?: return null
@@ -85,20 +73,18 @@ class PerNsHashMap<K : Any>(
         masks[index] = mask
     }
 
+    override fun createNewArrays(capacity: Int) {
+        items = arrayOfNulls(capacity)
+        masks = ByteArray(capacity)
+    }
+
     override fun rehash(newCapacity: Int) {
         val oldItems = items
         val oldMasks = masks
-        rehashTemplate(
-            createNewArrays = {
-                _set = arrayOfNulls(newCapacity)
-                items = arrayOfNulls(newCapacity)
-                masks = ByteArray(newCapacity)
-            },
-            moveValue = { newIndex, oldIndex ->
-                items[newIndex] = oldItems[oldIndex]
-                masks[newIndex] = oldMasks[oldIndex]
-            }
-        )
+        rehashTemplate(newCapacity) { newIndex, oldIndex ->
+            items[newIndex] = oldItems[oldIndex]
+            masks[newIndex] = oldMasks[oldIndex]
+        }
     }
 
     override val size: Int
@@ -177,175 +163,4 @@ private fun PerNs.asSingleVisItem(): Pair<VisItem, Namespace>? {
     values.singleOrNull()?.let { return it to Namespace.Values }
     macros.singleOrNull()?.let { return it to Namespace.Macros }
     return null
-}
-
-/**
- * Copy of [THashMap], abstracted over values array.
- * Deletion is not supported (we don't need it).
- */
-@Suppress("UNCHECKED_CAST")
-abstract class PerNsHashMapBase<K : Any, V : Any> : TObjectHash<K>(), MutableMap<K, V> {
-
-    abstract fun getValueAtIndex(index: Int): V?
-    abstract fun setValueAtIndex(index: Int, value: V)
-
-    override val size: Int
-        get() = _size
-
-    override fun put(key: K, value: V): V? {
-        var previous: V? = null
-        var index = insertionIndex(key)
-        val alreadyStored = index < 0
-        if (alreadyStored) {
-            index = -index - 1
-            previous = getValueAtIndex(index)
-        }
-        _set[index] = key
-        setValueAtIndex(index, value)
-        if (!alreadyStored) {
-            postInsertHook()
-        }
-        return previous
-    }
-
-    override fun putIfAbsent(key: K, value: V): V? {
-        val index = insertionIndex(key)
-        val alreadyStored = index < 0
-        if (alreadyStored) return getValueAtIndex(-index - 1)
-
-        _set[index] = key
-        setValueAtIndex(index, value)
-        postInsertHook()
-        return null
-    }
-
-    /** Key deletion is not supported, that's why we `usedFreeSlot` is always true */
-    private fun postInsertHook() = postInsertHook(true)
-
-    protected inline fun rehashTemplate(createNewArrays: () -> Unit, moveValue: (Int, Int) -> Unit) {
-        val oldCapacity = _set.size
-        val oldKeys = _set
-        createNewArrays()
-        var i = oldCapacity
-        while (i-- > 0) {
-            if (oldKeys[i] != null) {
-                val oldKey = oldKeys[i] as K
-                val index = insertionIndex(oldKey)
-                if (index < 0) {
-                    throwObjectContractViolation(_set[-index - 1], oldKey)
-                }
-                _set[index] = oldKey
-                moveValue(index, i)
-            }
-        }
-    }
-
-    override operator fun get(key: K): V? {
-        val index = index(key)
-        return if (index < 0) null else getValueAtIndex(index)
-    }
-
-    /** Unlike default implementation, doesn't call [containsKey] */
-    override fun getOrDefault(key: K, defaultValue: V): V = get(key) ?: defaultValue
-
-    override fun clear() = throw UnsupportedOperationException()
-
-    override fun remove(key: K): V = throw UnsupportedOperationException()
-
-    override fun removeAt(index: Int) = throw UnsupportedOperationException()
-
-    override val values: MutableCollection<V>
-        get() = ValueView()
-
-    override val keys: MutableSet<K>
-        get() = KeyView()
-
-    override val entries: MutableSet<MutableMap.MutableEntry<K, V>>
-        get() = EntryView()
-
-    override fun containsValue(value: V): Boolean = throw UnsupportedOperationException()
-
-    override fun containsKey(key: K): Boolean = contains(key)
-
-    override fun putAll(from: Map<out K, V>) {
-        ensureCapacity(from.size)
-        for ((key, value) in from) {
-            put(key, value)
-        }
-    }
-
-    private inner class KeyView : AbstractSet<K>() {
-        override fun iterator(): MutableIterator<K> {
-            return object : THashIterator<K>() {
-                override fun objectAtIndex(index: Int): K = _set[index] as K
-            }
-        }
-
-        override val size: Int get() = _size
-
-        override fun contains(element: K): Boolean = this@PerNsHashMapBase.contains(element)
-    }
-
-    private inner class ValueView : AbstractCollection<V>() {
-        override fun iterator(): MutableIterator<V> {
-            return object : THashIterator<V>() {
-                override fun objectAtIndex(index: Int): V = getValueAtIndex(index)!!
-            }
-        }
-
-        override val size: Int get() = _size
-
-        override fun contains(element: V): Boolean = throw UnsupportedOperationException()
-    }
-
-    private inner class EntryView : AbstractSet<MutableMap.MutableEntry<K, V>>() {
-        override fun iterator(): MutableIterator<Entry<K, V>> {
-            return object : THashIterator<Entry<K, V>>() {
-                override fun objectAtIndex(index: Int): Entry<K, V> {
-                    return Entry(_set[index] as K, getValueAtIndex(index)!!)
-                }
-            }
-        }
-
-        override val size: Int get() = _size
-
-        override fun contains(element: MutableMap.MutableEntry<K, V>): Boolean = throw UnsupportedOperationException()
-    }
-
-    private class Entry<K : Any, V : Any>(override val key: K, override val value: V) : MutableMap.MutableEntry<K, V> {
-        override fun setValue(newValue: V): V = throw UnsupportedOperationException()
-    }
-
-    private abstract inner class THashIterator<V> : MutableIterator<V> {
-
-        private var index: Int = capacity()
-
-        /**
-         * Returns the object at the specified index.
-         * Subclasses should implement this to return the appropriate object for the given index.
-         */
-        protected abstract fun objectAtIndex(index: Int): V
-
-        override operator fun hasNext(): Boolean = nextIndex() >= 0
-
-        override fun next(): V {
-            index = nextIndex()
-            if (index < 0) throw NoSuchElementException()
-            return objectAtIndex(index)
-        }
-
-        /**
-         * Returns the index of the next value in the data structure
-         * or a negative value if the iterator is exhausted.
-         */
-        private fun nextIndex(): Int {
-            val set = _set
-            var i = index
-            @Suppress("ControlFlowWithEmptyBody")
-            while (i-- > 0 && set[i] == null);
-            return i
-        }
-
-        override fun remove() = throw UnsupportedOperationException()
-    }
 }
